cleaned up some files and moved onEnter functions to separate file
